Fixed peer inequality check in Python 2.7
In Python 2.7, there is a slightly different behaviour of the not equal operation (!=) compared to Python 3. In particular, in Python 2.7, statements like peerA == peerB would evaluate to True but statements like peerA != peerB would not evaluate to False, due to the absence of a __ne__ method in the Peer class. This PR fixes this issue and modifies an existing test to test this change.